requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.20.0
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.7.0
attrs==24.3.0
beautifulsoup4==4.12.3
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Crawl4AI==0.4.246
cryptography==44.0.0
distro==1.9.0
fake-http-header==0.3.5
fastapi==0.115.6
filelock==3.16.1
firebase-admin==6.6.0
frozenlist==1.5.0
fsspec==2024.12.0
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.27.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.56.6
lxml==5.3.0
MarkupSafe==3.0.2
mockito==1.5.3
msgpack==1.1.0
multidict==6.1.0
nltk==3.9.1
numpy==1.26.0
openai==1.58.1
packaging==24.2
pillow==10.4.0
playwright==1.49.1
pluggy==1.5.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyee==12.0.0
PyJWT==2.10.1
pyOpenSSL==24.3.0
pyparsing==3.2.1
pytest==8.3.4
pytest-mockito==0.0.4
python-dotenv==1.0.1
PyYAML==6.0.2
rank-bm25==0.2.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rpds-py==0.22.3
rsa==4.9
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
starlette==0.41.3
tavily-python==0.5.0
tf-playwright-stealth==1.1.0
tiktoken==0.8.0
tokenizers==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
